Create README.md
Smart-Home-Automation
This project explains the working of smart home system, implemented through 'tinkercad' simulator. This system contains automatic switching of fan or air-conditioner depending on the room temperature, automatic opening of doors and motion sensor light for stairs.

click smart home codes.ino to view project code.

link to my tinkercad project: link to my tinkercad project

Working: change the value of temperature sensor to above 35 degree to switch on motor(fan).

change the distance of ultrasonic sensor below 100 cm to turn on motor to open door.

move the motion sensor to switch on the light. It is used in the staircase.
This project explains the working of smart home system, implemented through 'tinkercad' simulator. This system contains automatic switching of fan or air-conditioner depending on the room temperature, automatic opening of doors and motion sensor light for stairs.

click smart home codes.ino to view project code.

link to my tinkercad project: link to my tinkercad project

Working: change the value of temperature sensor to above 35 degree to switch on motor(fan).

change the distance of ultrasonic sensor below 100 cm to turn on motor to open door.

move the motion sensor to switch on the light. It is used in the staircase.

buzzer sensor active when gas sensor above 200